new grapple system implemented
Now instead of the player sending raycasts to the grapple points, the grapple points now keep track of where the player, if they are in range then they send a raycast to see if there is line of sight. Then it returns itself (gameObject) if the input is within the aim assist tolerance.
This new system now also keeps all currently active grapple points in a list and iterates through them when the grapple input is released in order to see if the player is pointing towards one of them.