bug fix
There seems to be a bug where if the grapple speed is too fast and the grapple stop distance too low, the velocity flips right at the end of the grapple reversing all speed.
This is due to the player overshooting the grapple point before it can check if it should detach the grapple.