bug fix: gravity during slow down
The gravity was being scaled by the delta time not the fixed delta timeGravity wasn't being correctly set during grapple aiming, and if the gravity was the grounding force when the player starts to aim, the player can float horizontally while aiming. This was fixed by just copying the gravity logic from the MoveState.