animations use event system now
Instead of having the various attack animations set a boolean during the script, they now have an event keyframe at the end which calls a method within the player controller.
There is also a better trigger system where the states just call one method to try and consume the trigger and if it's there then it sends true. This removes the responsibility
of reseting the trigger onto the player controller. With this new system, now the end animation event can be triggered on the last frame which fixes the bug where the player sprite can be flipped directions before finishing the attack animation.